Lab_of_the_Future_Workshop/app/streamlit-alternative/requirements.txt
streamlit>=1.28.0
databricks-sql-connector>=2.9.0
pandas>=2.0.0
plotly>=5.18.0
